echec du debug pour videos protégées
