+++ templates/instrument/index.html.twig
@@ -19,7 +19,7 @@
         {% for instrument in instruments %}
             <div class="p-4 border rounded flex flex-col items-center bg-white shadow">
                 {% if instrument.pictureUrl %}
-                    <img src="{{ instrument.pictureUrl }}" alt="{{ instrument.nameInstr }}" width="180" height="180" class="w-60 h-60 object-fit">
+                    <img src="{{ instrument.pictureUrl }}" alt="{{ instrument.nameInstr }}" class="w-80 h-80 object-cover">
                     {# on ajoute explcitement la taille 
                      #}
                 {% endif %}
